dev: update bulk import endpoint
