+++ apiserver/plane/api/views/issue.py
@@ -4,7 +4,7 @@ import random
 from itertools import groupby, chain
 
 # Django imports
-from django.db.models import Prefetch, OuterRef, Func, F, Q, Max
+from django.db.models import Prefetch, OuterRef, Func, F, Q
 from django.core.serializers.json import DjangoJSONEncoder
 
 # Third Party imports
@@ -43,13 +43,9 @@ from plane.db.models import (
     CycleIssue,
     ModuleIssue,
     IssueLink,
-    State,
-    IssueSequence,
-    IssueLabel,
 )
 from plane.bgtasks.issue_activites_task import issue_activity
 from plane.utils.grouper import group_results
-from plane.utils.html_processor import strip_tags
 
 
 class IssueViewSet(BaseViewSet):
@@ -722,177 +718,6 @@ class IssueLinkViewSet(BaseViewSet):
             .filter(project__project_projectmember__member=self.request.user)
             .distinct()
         )
-
-
-class BulkCreateIssuesEndpoint(BaseAPIView):
-    def post(self, request, slug, project_id):
-        try:
-            # Get the project
-            project = Project.objects.get(pk=project_id, workspace__slug=slug)
-
-            # Get the default state
-            default_state = State.objects.filter(
-                project_id=project_id, default=True
-            ).first()
-            # if there is no default state assign any random state
-            if default_state is None:
-                default_state = State.objects.filter(project_id=project_id).first()
-
-            # Get the maximum sequence_id
-            last_id = IssueSequence.objects.filter(project_id=project_id).aggregate(
-                largest=Max("sequence")
-            )["largest"]
-
-            last_id = 1 if last_id is None else last_id + 1
-
-            # Get the maximum sort order
-            largest_sort_order = Issue.objects.filter(
-                project_id=project_id, state=default_state
-            ).aggregate(largest=Max("sort_order"))["largest"]
-
-            largest_sort_order = (
-                65535 if largest_sort_order is None else largest_sort_order + 10000
-            )
-
-            # Get the issues_data
-            issues_data = request.data.get("issues_data", [])
-
-            if not len(issues_data):
-                return Response(
-                    {"error": "Issue data is required"},
-                    status=status.HTTP_400_BAD_REQUEST,
-                )
-
-            # Issues
-            bulk_issues = []
-            for issue_data in issues_data:
-                bulk_issues.append(
-                    Issue(
-                        project_id=project_id,
-                        workspace_id=project.workspace_id,
-                        state=default_state,
-                        name=issue_data.get("name", "Issue Created through Bulk"),
-                        description_html=issue_data.get("description_html", "<p></p>"),
-                        description_stripped=(
-                            None
-                            if (
-                                issue_data.get("description_html") == ""
-                                or issue_data.get("description_html") is None
-                            )
-                            else strip_tags(issue_data.get("description_html"))
-                        ),
-                        sequence_id=last_id,
-                        sort_order=largest_sort_order,
-                        start_date=issue_data.get("start_date", None),
-                        target_date=issue_data.get("target_date", None),
-                    )
-                )
-
-                largest_sort_order = largest_sort_order + 10000
-                last_id = last_id + 1
-
-            issues = Issue.objects.bulk_create(
-                bulk_issues,
-                batch_size=100,
-                ignore_conflicts=True,
-            )
-
-            # Sequences
-            _ = IssueSequence.objects.bulk_create(
-                [
-                    IssueSequence(
-                        issue=issue,
-                        sequence=issue.sequence_id,
-                        project_id=project_id,
-                        workspace_id=project.workspace_id,
-                    )
-                    for issue in issues
-                ],
-                batch_size=100,
-            )
-
-            # Attach Labels
-            bulk_issue_labels = []
-            for issue, issue_data in zip(issues, issues_data):
-                labels_list = issue_data.get("labels_list", [])
-                bulk_issue_labels = bulk_issue_labels + [
-                    IssueLabel(
-                        issue=issue,
-                        label_id=label_id,
-                        project_id=project_id,
-                        workspace_id=project.workspace_id,
-                        created_by=request.user,
-                        updated_by=request.user,
-                    )
-                    for label_id in labels_list
-                ]
-
-            _ = IssueLabel.objects.bulk_create(bulk_issue_labels, batch_size=100)
-
-            # Create Comments
-            bulk_issue_comments = []
-            for issue, issue_data in zip(issues, issues_data):
-                comments_list = issue_data.get("comments_list", [])
-                bulk_issue_comments = bulk_issue_comments + [
-                    IssueComment(
-                        issue=issue,
-                        comment_html=comment.get("comment_html", "<p></p>"),
-                        actor=request.user,
-                        project_id=project_id,
-                        workspace_id=project.workspace_id,
-                        created_by=request.user,
-                        updated_by=request.user,
-                    )
-                    for comment in comments_list
-                ]
-
-            _ = IssueComment.objects.bulk_create(bulk_issue_comments, batch_size=100)
-
-            # Attach Links
-            _ = IssueLink.objects.bulk_create(
-                [
-                    IssueLink(
-                        issue=issue,
-                        url=issue_data.get("link", {}).get("url", "https://github.com"),
-                        title=issue_data.get("link", {}).get("title", "Original Issue"),
-                        project_id=project_id,
-                        workspace_id=project.workspace_id,
-                        created_by=request.user,
-                        updated_by=request.user,
-                    )
-                    for issue, issue_data in zip(issues, issues_data)
-                ]
-            )
-
-            # Track the issue activities
-            [
-                issue_activity.delay(
-                    {
-                        "type": "issue.activity.created",
-                        "requested_data": None,
-                        "actor_id": str(request.user.id),
-                        "issue_id": str(issue.id),
-                        "project_id": str(project_id),
-                        "current_instance": None,
-                    },
-                )
-                for issue in issues
-            ]
-
-            return Response(
-                {"issues": IssueFlatSerializer(issues, many=True).data},
-                status=status.HTTP_201_CREATED,
-            )
-        except Project.DoesNotExist:
-            return Response(
-                {"error": "Project Does not exist"}, status=status.HTTP_404_NOT_FOUND
-            )
-        except Exception as e:
-            capture_exception(e)
-            return Response(
-                {"error": "Something went wrong please try again later"},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
 
 
 class BulkCreateIssueLabelsEndpoint(BaseAPIView):
